% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Erl}
\alias{Erl}
\title{Erlang}
\description{
Interact with the erlang shell.
}
\examples{

## ------------------------------------------------
## Method `Erl$eval`
## ------------------------------------------------

Erl$new()$eval("3 + 4.")

## ------------------------------------------------
## Method `Erl$assign`
## ------------------------------------------------

df <- data.frame(
  key = letters[1:10],
  value = runif(10)
)

map <- as_map(df, key = key, value = value)

e <- Erl$new()
e$assign("Vehicles", map)
e$eval("maps::get(a, Vehicles)")
e$eval("maps:update(b, 25, Vehicles)")

## ------------------------------------------------
## Method `Erl$compile`
## ------------------------------------------------

\dontrun{Erl$new()$compile("test")}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Erl$new()}}
\item \href{#method-eval}{\code{Erl$eval()}}
\item \href{#method-assign}{\code{Erl$assign()}}
\item \href{#method-compile}{\code{Erl$compile()}}
\item \href{#method-finalize}{\code{Erl$finalize()}}
\item \href{#method-halt}{\code{Erl$halt()}}
\item \href{#method-print}{\code{Erl$print()}}
\item \href{#method-clone}{\code{Erl$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Erl$new(bin = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bin}}{Path to erlang binary.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Initialise Class
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eval"></a>}}
\subsection{Method \code{eval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Erl$eval(code, wait = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{Code to evaluate.}

\item{\code{wait}}{Whether to re-attempt to evaluate if at first fails.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Evaluate Erlang Code
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{Erl$new()$eval("3 + 4.")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-assign"></a>}}
\subsection{Method \code{assign()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Erl$assign(name, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of variable.}

\item{\code{value}}{Value of variable, an object of class \code{erlang_object}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a variable
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{df <- data.frame(
  key = letters[1:10],
  value = runif(10)
)

map <- as_map(df, key = key, value = value)

e <- Erl$new()
e$assign("Vehicles", map)
e$eval("maps::get(a, Vehicles)")
e$eval("maps:update(b, 25, Vehicles)")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-compile"></a>}}
\subsection{Method \code{compile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Erl$compile(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Name of file to compile, passed to \code{c}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Compile Erlang
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{Erl$new()$compile("test")}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Erl$finalize()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Kills the session.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-halt"></a>}}
\subsection{Method \code{halt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Erl$halt()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Kills the session.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Erl$print()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Print the session.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Erl$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
