% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{erlang_object}
\alias{erlang_object}
\title{Erlang Data Structure}
\usage{
erlang_object(x, cl = NULL)
}
\arguments{
\item{x}{The data.}

\item{cl}{The class name}
}
\description{
Erland Data Struture class for better printing 
and potential future improvements.
}
\keyword{internal}
