% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{map}
\alias{map}
\alias{as_map}
\alias{as_map.default}
\alias{as_map.list}
\alias{as_map.data.frame}
\title{Convert R Objects to Erlang Map.}
\usage{
as_map(data, ...)

\method{as_map}{default}(data, ...)

\method{as_map}{list}(data, ...)

\method{as_map}{data.frame}(data, ..., key, value)
}
\arguments{
\item{data}{A data.frame or list to convert to a map.}

\item{...}{Options passed to methods.}

\item{key, value}{Bare columns names of key-value pair columns.}
}
\description{
Convert R objects to Erlang a key-value assocuation map.
}
\examples{
as_map(c("a", 1))

# from a data.frame
df <- data.frame(
  key = letters[1:10],
  value = sample(1:10, 10)
)

as_map(df, key = key, value = value)

# from a list
lst <- list(
  list("a", 1),
  list("b", 2)
)

as_map(lst)

}
