% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{tuple}
\alias{tuple}
\alias{as_tuple}
\alias{as_list}
\title{Convert R Objects to Tuples or Lists}
\usage{
as_tuple(data)

as_list(data)
}
\arguments{
\item{data}{The data to convert.}
}
\description{
Convert R objects to tuples or lists.
}
\examples{
as_tuple(cars)

}
